feat(capture-sdk): Move core ui compose components to capture-sdk module
PP-403
