+++ capture-sdk/sdk/src/main/java/net/gini/android/capture/ui/components/switcher/GiniSwitchColors.kt
@@ -1,9 +1,9 @@
-package net.gini.android.bank.sdk.capture.skonto.components.switcher
+package net.gini.android.capture.ui.components.switcher
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.Immutable
 import androidx.compose.ui.graphics.Color
-import net.gini.android.bank.sdk.ui.theme.GiniTheme
+import net.gini.android.capture.ui.theme.GiniTheme
 
 @Immutable
 data class GiniSwitchColors(
